Change lowercase? and uppercase? for non-alphabetics
This is a backwards-incompatible change and will be followed by a commit to
increment the major version of the strings package.

It's unintuitive for `lowercase?('8')` to return `#t` since being lowercase
implies being alphabetic, and similarly for `uppercase?`. So this change makes
both `lowercase?('8')` and `uppercase?('8')` return `#f`.

What exactly `lowercase?("f8")` should mean is less clear, but as a member of
the generic function it should match the method on `<character>` which, I
believe, implies that it should be equivalent to `every?(lowercase?, string)`.
Because the implementation is trivial and because I can't think of a common use
case for this method, I'm removing it entirely and making this a non-generic
function.